refactor(Test): domainTest, RepositoryTest 로 나눔
domainTest, RepositoryTest 로 나눔
