+++ src/test/java/qna/domain/UserTest.java
@@ -8,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import java.time.LocalDateTime;
 import java.util.NoSuchElementException;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -17,27 +16,18 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @DataJpaTest
 public class UserTest {
-    public static final User JAVAJIGI = new User(1L, "javajigi", "password", "name", "javajigi@slipp.net");
-    public static final User SANJIGI = new User(2L, "sanjigi", "password", "name", "sanjigi@slipp.net");
-    public static final User LEWISSEO = new User(3L, "lewisseo", "password", "name", "lewisseo91@random.net");
+    public static final User JAVAJIGI = new User("javajigi", "password", "name", "javajigi@slipp.net");
+    public static final User SANJIGI = new User("sanjigi", "password", "name", "sanjigi@slipp.net");
+    public static final User LEWISSEO = new User("lewisseo", "password", "name", "lewisseo91@random.net");
 
     @Autowired
     private UserRepository userRepository;
 
-    @Autowired
-    private QuestionRepository questionRepository;
-
     private User user;
-
-    private Question question;
-
-    private LocalDateTime now;
 
     @BeforeEach
     void setup() {
-        now = LocalDateTime.now();
         user = userRepository.save(JAVAJIGI);
-        question = questionRepository.save(new Question("ex title", "ex contents"));
     }
 
     @DisplayName("user 생성")
@@ -49,12 +39,6 @@ public class UserTest {
         );
     }
 
-    @DisplayName("user userId로 찾기")
-    @Test
-    void findUserByIdTest() {
-        assertThat(userRepository.findByUserId(new UserId("javajigi"))).isNotNull();
-    }
-
     @DisplayName("user 수정")
     @Test
     void userUpdateTest() {
@@ -62,14 +46,6 @@ public class UserTest {
         User userFromRepository = userRepository.findById(user.getId())
                 .orElseThrow(NoSuchElementException::new);
         assertThat(userFromRepository.getEmail()).isEqualTo("change_mail@slipp.net");
-    }
-
-    @DisplayName("user 삭제")
-    @Test
-    void removeUserTest() {
-        assertThat(userRepository.findAll().size()).isEqualTo(1);
-        userRepository.delete(user);
-        assertThat(userRepository.findAll().size()).isZero();
     }
 
     @AfterEach
